Centralize probability blending
