+++ core/market_pricer.py
@@ -156,16 +156,6 @@ def compute_moneyline(home_scores, away_scores):
     }
 
 
-def logistic_decay(t_hours, t_switch=8, slope=1.5):
-    return 1 / (1 + math.exp((t_switch - t_hours) / slope))
-
-def base_model_weight_for_market(market):
-    if "1st" in market:
-        return 0.9  # prioritize derivatives (1st innings) first
-    elif market.startswith("h2h") or (market.startswith("spreads") and "_" not in market) or (market.startswith("totals") and "_" not in market):
-        return 0.6  # mainlines (h2h, spreads, totals without "_")
-    else:
-        return 0.75  # fallback for anything else
 
 def get_market_price(market_dict, market, side):
     """
@@ -205,20 +195,6 @@ def get_market_price(market_dict, market, side):
 
 
 
-def blend_prob(p_model, market_odds, market_type, hours_to_game, p_market=None):
-    """
-    Blend model simulation probability with market consensus probability.
-    """
-    if p_market is None:
-        p_market = implied_prob(market_odds)
-
-    base_weight = base_model_weight_for_market(market_type)
-    w_time = logistic_decay(hours_to_game, t_switch=8, slope=1.5)
-    w_model = min(base_weight * w_time, 1.0)
-    w_market = 1 - w_model
-
-    p_blended = w_model * p_model + w_market * p_market
-    return p_blended, w_model, p_model, p_market
 
 
 # === CLI Summary Formatter ===
